Tables & API sets
Moving two utilities that are each used in multiple places to their own files: tables and the API sets
