+++ commands/submissions.go
@@ -6,7 +6,6 @@ import (
 	"strings"
 
 	"github.com/spf13/cobra"
-	interactor "github.com/tuupke/api-interactor"
 )
 
 var submissionsCommand = &cobra.Command{
@@ -14,43 +13,6 @@ var submissionsCommand = &cobra.Command{
 	Short:   "Shows past submissions and their judgements",
 	RunE:    submissions,
 	PreRunE: configHelper("baseurl"),
-}
-
-type (
-	judgementTypeSet []interactor.JudgementType
-	judgementSet     []interactor.Judgement
-	teamSet          []interactor.Team
-)
-
-func (j judgementTypeSet) byId(id string) (interactor.JudgementType, bool) {
-	for _, jt := range j {
-		if jt.Id == id {
-			return jt, true
-		}
-	}
-
-	return interactor.JudgementType{}, false
-}
-
-func (t teamSet) byId(id string) (interactor.Team, bool) {
-	for _, team := range t {
-		if team.Id == id {
-			return team, true
-		}
-	}
-
-	return interactor.Team{}, false
-}
-
-func (j judgementSet) bySubmissionId(id string) (judgementSet, bool) {
-	var jud []interactor.Judgement
-	for _, ju := range j {
-		if ju.SubmissionId == id {
-			jud = append(jud, ju)
-		}
-	}
-
-	return jud, len(jud) > 0
 }
 
 func submissions(cmd *cobra.Command, args []string) error {
